// src/types/raw.ts
export type CommonOptionName = "home_default" | "home_custom" | (string & {});

export type CommonImage = {
  option_name?: CommonOptionName;
  img_url?: string;
  crop_ratio?: string;
};

export type CommonHighlightVideo = {
  content_id?: string;
  dash_url?: string;
  hls_url?: string;
};

export type CommonContentRating =
  | "전체 이용가"
  | "7세 이용가"
  | "12세 이용가"
  | "15세 이용가"
  | "성인 이용가"
  | (string & {});

export type CommonRatingType =
  | "방송통신심의위원회"
  | "영상물등급위원회"
  | (string & {});

export type CommonMedium = "TVA" | "극장판" | "OVA" | "기타" | (string & {});

export type CommonRatingComponent =
  | "선정성"
  | "주제"
  | "폭력성"
  | "모방위험"
  | "대사"
  | "공포"
  | "약물"
  | (string & {});
// KCC/영등위 상 가능성 있음: 차별, 혐오, 도박, 범죄, 음주, 흡연, 담배, 충격, 언어, 음란성, 유해정보

export type MaxEpisodeRating = {
  rating?: number | unknown;
  rating_type?: CommonRatingType | unknown;
  classification_number?: string;
  broadcast_channel_name?: string;
  broadcast_date?: Date | unknown;
  rating_components?: CommonRatingComponent[] | unknown[];
};

export type CommonPerson = { name?: string; role?: string };
export type CommonCast = {
  character_name?: string;
  voice_actor_names?: string[];
};
export type CommonProductionCompany = { name?: string };

export type Version = {
  current?: string;
  required?: string;
  required_build_number?: number;
};

export type ProfileRank = {
  rank?: number;
  continued_membership_days?: number | unknown;
};

export type ProfilesV1MyProfile =
  | {
      id: number;
      account_id: number;
      name: string;
      image: string;
      status: string;
      profile_rank: ProfileRank;
      is_locked: boolean;
      is_for_kids: boolean;
      content_rating: number;
      is_default: boolean;
      is_main: boolean;
    }
  | WTF;
// ...
export type ItemsV4ID = {
  id: number;
  uid: string;
  name: string;
  content: string;
  is_adult: boolean;
  is_uncensored: boolean;
  is_dubbed: boolean;
  is_laftel_only: boolean;
  is_laftel_original: boolean;
  is_ending: boolean;
  is_exclusive: boolean;
  is_avod: boolean;
  is_svod: boolean;
  is_new_release: boolean;
  is_upcoming_release: boolean;
  is_episode_existed: boolean;
  medium: CommonMedium;
  images: CommonImage[];
  genre: string[];
  release_weekdays: string[];
  latest_episode_release_datetime: string;
  avg_rating: number;
  logo_img: string | null;
  color_code: string | null;
  description: string | null;
  is_viewing: boolean;
  highlight_video: CommonHighlightVideo | null;
  awards: string[];
  notice: string;
  tags: string[];
  production: string | null;
  air_year_quarter: string | null;
  copyright: string;
  author: unknown[];
  illustrator: unknown[];
  expire_datetime: string | null;
  series_id: number | null;
  max_episode_rating: MaxEpisodeRating;
  directors: CommonPerson[];
  casts: CommonCast[];
  production_companies: CommonProductionCompany[];
  simulcast_channel: unknown | null;
};

export type ItemList = ItemsV4ID;

export type ItemsV1ScheduledExpiredItem = PaginatedResponse<{
  expire_date_with_hour?: Date | string;
  items?: Array<{
    id?: number;
    name?: string;
    images?: CommonImage[];
  }>;
}>;

export type ItemsV2SeriesID = PaginatedResponse<{
  id: number;
  name: string;
  img: string;
  content_rating: CommonContentRating;
  images: CommonImage[];
}>;

export type EpisodeProduct = {
  id?: number;
  name?: string;
  list_price?: number;
  period?: string;
  promotion?: null | unknown;
};

export type EpisodesV3 = {
  id: number;
  title: string;
  subject: string;
  description: string;
  episode_num: string;
  episode_order: number;
  thumbnail_path: string | null;
  has_preview: boolean;
  item_expire_datetime: string | null;
  in_app_download: boolean;
  is_avod: boolean;
  is_free: boolean;
  is_viewing: boolean;
  published_datetime: string | null;
  running_time: string;
  progressbar: null | number;
  episode_products: EpisodeProduct[];
  rating: MaxEpisodeRating;
  access_info_list: unknown[];
  access_type: string;
  is_final: boolean;
};
// ...
export type CommentItem = {
  id?: number;
  parent_comment_id?: number | null;
  profile?: ProfilesV1MyProfile;
  content?: string;
  item?: Partial<ItemsV4ID>;
  episode?: {
    id?: number;
    subject?: string;
    episode_num?: string;
    thumbnail_path?: string | null;
  };
  created?: Date | unknown;
  modified?: Date | unknown;
  count_like?: number;
  count_reply_comment?: number;
  is_click_like?: boolean;
  is_spoiler?: boolean;
};
// ...
export type HomeV1RecommendRanking = {
  id?: number;
  name?: string;
  img?: string;
  cropped_img?: string;
  is_adult?: boolean;
  images?: CommonImage[];
  genres?: string[];
  highlight_video?: CommonHighlightVideo | null;
  is_laftel_only?: boolean;
  is_laftel_original?: boolean;
  is_uncensored?: boolean;
  is_dubbed?: boolean;
  is_avod?: boolean;
  is_viewing?: boolean;
  content_rating?: CommonContentRating;
  rating?: number;
  rating_type?: CommonRatingType;
  medium?: CommonMedium;
  is_ending?: boolean;
};
// ...
export type WTF =
  | {
      detail?: string | unknown;
      error?: unknown;
      code?: string | unknown;
      data?: null | unknown;
      msg?: string | unknown;
      ip?: string | unknown;
      country_code?: string | unknown;
    }
  | { detail: "재생 가능 시간 초과"; code: "PLAYBACK_EXPIRED" }
  | {
      detail: "대한민국 이외 지역에서는\n저작권 문제로 시청할 수 없습니다.";
      code: "BLOCKED_BY_DISALLOWED_ACCESS";
      data: null;
    }
  | {
      code?: "PLAYBACK_DEACTIVATED" | string;
      detail?: Array<{
        user_id: number;
        profile_name: string;
        item_name: string;
        device_type: string;
      }>;
    }
  | {
      detail: "episode_ids, item_id 중에 하나는 필수 값 입니다.";
      code: "INVALID";
    };

export type AuthResponse = {
  user: ProfilesV1MyAccount;
  key: string;
  method: "email" | "google" | "kakao" | string;
  is_restored: boolean;
  is_registered: boolean;
};

export type AccountsPasswordReset = {
  msg: string;
  email: string;
  expire_datetime: string;
};

export type ProductsV3MembershipProducts = {
  products: MembershipProduct[];
};

export type PlaybackInfo = {
  op_start?: number | null;
  op_end?: number | null;
  ed_start?: number | null;
  ed_end?: number | null;
  episode_id?: number;
  episode_num?: string;
  action_time?: number;
  progressbar?: unknown | null;
  access_type?: string;
  episode_type?: string;
};

export type ProductsInfo = {
  point?: number;
  products?: Array<{
    product_no?: string;
    name?: string;
    is_rental?: boolean;
    price?: number;
  }>;
};

export type ProtectedStreamingInfo = {
  content_id?: string;
  access_type?: string;
  widevine_token?: string | null;
  fairplay_token?: string | null;
  playready_token?: string | null;
  dash_url?: string;
  hls_url?: string;
  subtitle_url?: unknown;
};

export type PublicStreamingInfo = {
  dash_preview_url?: string | null;
  hls_preview_url?: string | null;
  thumbnail?: string | null;
  subtitle_preview_url?: string | null;
};

export type StreamingInfoV2 =
  | {
      is_cartoon_network?: boolean;
      playback_info?: PlaybackInfo;
      products_info?: ProductsInfo;
      protected_streaming_info?: ProtectedStreamingInfo;
      public_streaming_info?: PublicStreamingInfo;
      play_log_id?: number;
    }
  | {
      code?: "PLAYBACK_DEACTIVATED" | string;
      detail?: Array<{
        user_id: number;
        profile_name: string;
        item_name: string;
        device_type: string;
      }>;
    }
  | WTF;
// ...
export type ReviewsV1MyReview = {
  id?: null | number;
  is_click_like?: boolean;
  count_like?: number;
  profile?: ProfilesV1MyProfile;
  content?: string;
  is_spoiler?: boolean;
  score?: number;
  item?: null | number;
  created?: null | Date;
  modified?: null | Date;
};